Replace stubs with real modules and update tests
