+++ tests/test_bootstrap_initialization.py
@@ -1,14 +1,6 @@
 import unittest
-from tests.helpers import stubs
-
-# Provide missing PrimeInstruction for Gödel loop imports if absent
-import core.instruction_set as _instruction_set
-if not hasattr(_instruction_set, "PrimeInstruction"):
-    _instruction_set.PrimeInstruction = stubs.PrimeInstruction
-
-import builtins
-if not hasattr(builtins, "StateTransitionManager"):
-    builtins.StateTransitionManager = stubs.StateTransitionManager
+from core.instruction_set import PrimeInstruction
+from consciousness.state_transition_manager import StateTransitionManager
 
 from core.prime_vm import ConsciousPrimeVM
 from consciousness.consciousness_integration import ConsciousnessIntegrator
