+++ tests/test_consciousness_ecosystem.py
@@ -9,7 +9,6 @@ import pytest
 import asyncio
 import numpy as np
 import networkx as nx
-from tests.helpers import stubs
 from unittest.mock import Mock, patch, AsyncMock
 from datetime import datetime
