+++ tests/test_error_logging.py
@@ -2,15 +2,8 @@ import unittest
 from unittest import mock
 import networkx
 import numpy as np
-from tests.helpers import stubs
-
-import core.instruction_set as _instruction_set
-if not hasattr(_instruction_set, "PrimeInstruction"):
-    _instruction_set.PrimeInstruction = stubs.PrimeInstruction
-
-import builtins
-if not hasattr(builtins, "StateTransitionManager"):
-    builtins.StateTransitionManager = stubs.StateTransitionManager
+from core.instruction_set import PrimeInstruction
+from consciousness.state_transition_manager import StateTransitionManager
 
 from consciousness.consciousness_integration import Condition
 from modules.strange_loops.loop_detector import StrangeLoopDetector
