+++ tests/test_harm_detection.py
@@ -2,15 +2,8 @@ import pytest
 import asyncio
 import numpy
 import networkx
-from tests.helpers import stubs
-
-import core.instruction_set as _instruction_set
-if not hasattr(_instruction_set, "PrimeInstruction"):
-    _instruction_set.PrimeInstruction = stubs.PrimeInstruction
-
-import builtins
-if not hasattr(builtins, "StateTransitionManager"):
-    builtins.StateTransitionManager = stubs.StateTransitionManager
+from core.instruction_set import PrimeInstruction
+from consciousness.state_transition_manager import StateTransitionManager
 
 
 from modules.consciousness_evolution.evolution_engine import (
