+++ tests/test_quantum_reality_helpers.py
@@ -3,7 +3,6 @@ import asyncio
 import numpy as np
 import scipy.constants  # noqa:F401
 import scipy.linalg  # noqa:F401
-from tests.helpers import stubs
 from modules.universe_interface.quantum_reality_interface import (
     QuantumRealityInterface,
     SpacetimeManipulation,
